+++ sdwan_policy_objects.tf
@@ -108,6 +108,16 @@ resource "sdwan_policy_object_policer" "policy_object_policer" {
   }]
 }
 
+resource "sdwan_policy_object_standard_community_list" "policy_object_standard_community_list" {
+  for_each           = { for p in try(local.feature_profiles.policy_object_profile.standard_community_lists, {}) : p.name => p }
+  name               = each.value.name
+  description        = try(each.value.description, null)
+  feature_profile_id = sdwan_policy_object_feature_profile.policy_object_feature_profile[0].id
+  entries = [for e in try(each.value.standard_communities, []) : {
+    standard_community = e
+  }]
+}
+
 resource "sdwan_policy_object_tloc_list" "policy_object_tloc_list" {
   for_each           = { for p in try(local.feature_profiles.policy_object_profile.tloc_lists, {}) : p.name => p }
   name               = each.value.name