+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - fix vty_line_logging parameter not being set in sdwan_system_global_feature
 - add support for sdwan_policy_object_as_path_list resource
+- add support for sdwan_policy_object_standard_community_list resource
 
 ## 1.0.0